# Improve Message.getBlob performance